Move and isolate changes to the StorefrontPreview component
